implemented retrieveUpcomingInvoice in service layer

+++ turbolessons/payment-service/src/main/java/com/turbolessons/paymentservice/dto/InvoiceData.java
@@ -1,6 +1,9 @@
 package com.turbolessons.paymentservice.dto;
 
-public record InvoiceData(String id, String account_name, String customer, int amount_due, int amount_paid,
-                          int amount_remaining, int attempt_count, boolean attempted, long created, long due_date,
+import com.stripe.model.Invoice;
+
+public record InvoiceData(String id, String account_name, String customer, long amount_due, long amount_paid,
+                          long amount_remaining, long attempt_count, boolean attempted, long created, long due_date,
                           long effective_at, long ending_balance, long next_payment_attempt, boolean paid) {
+
 }
